[refactor] 핸들러 생성자에 인터페이스 파라미터 설정
레포지토리 인터페이스 주입으로 생성자를 작성하였습니다.
DependencyFactory 에서 해당 인터페이스에 맞는 구현체를 넣어줄 것을 기대합니다.
